+++ src/robopilot/diff/__init__.py
@@ -1,0 +1,5 @@
+"""ProjectSpec diff helpers."""
+
+from robopilot.diff.spec_diff import SpecDiffResult, diff_spec_files, diff_specs
+
+__all__ = ["SpecDiffResult", "diff_spec_files", "diff_specs"]